TRAIL-34: Refactor: transfer Forms into RegistrationForm.tsx and Feat: Login Beautified
