+++ src/pages/LoginPage/index.tsx
@@ -101,50 +101,54 @@ export default function RegisterPage() {
     }
 
     return (
-        <div className='flex flex-col items-center'>
-            <h1>Register Page</h1>
-            <Form {...form}>
-                <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-6 w-1/6">
-                    <FormField
-                        control={form.control}
-                        name="email"
-                        render={({ field }) => (
-                            <FormItem className='mb-6'>
-                                <FormLabel>Email</FormLabel>
-                                <FormControl>
-                                    <Input placeholder="Email" {...field} />
-                                </FormControl>
-                            <FormMessage />
-                            </FormItem>
-                        )}
-                    />
-                    <FormField
-                        control={form.control}
-                        name="password"
-                        render={({ field }) => (
-                            <FormItem>
-                                <FormLabel>Password</FormLabel>
-                                <FormControl>
-                                    <div className='flex flex-col space-y-4'>
-                                        <Input type={showPassword ? "text" : "password"} placeholder="Password" {...field} />
-                                        <div className='items-top flex space-x-2'>
-                                            <Checkbox id='showPassword' onClick={() => setShowPassword(!showPassword)}> Show Password</Checkbox>
-                                            <label
-                                              htmlFor="showPassword"
-                                              className="text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70"
-                                            >
-                                              Show Password
-                                            </label>
-                                        </div>
-                                    </div>
-                                </FormControl>
+        <div className='flex flex-col min-h-screen justify-center items-center'>
+            <div className='pb-6'>
+                <h1 className='font-bold text-2xl'>Login your account</h1>
+            </div>
+            <div className='border rounded-xl w-[350px] shadow-lg dark:bg-zinc-900 bg-muted p-8'>
+                <Form {...form}>
+                    <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-6">
+                        <FormField
+                            control={form.control}
+                            name="email"
+                            render={({ field }) => (
+                                <FormItem className='mb-6'>
+                                    <FormLabel>Email</FormLabel>
+                                    <FormControl>
+                                        <Input placeholder="Email" {...field} />
+                                    </FormControl>
                                 <FormMessage />
-                            </FormItem>
-                        )}
-                    />
-                    <Button disabled={isLoading} type="submit">{isLoading && <Loader2 className='mr-2 h-4 w-4 animate-spin' />} Submit</Button>
-                </form>
-            </Form>
+                                </FormItem>
+                            )}
+                        />
+                        <FormField
+                            control={form.control}
+                            name="password"
+                            render={({ field }) => (
+                                <FormItem>
+                                    <FormLabel>Password</FormLabel>
+                                    <FormControl>
+                                        <div className='flex flex-col space-y-4'>
+                                            <Input type={showPassword ? "text" : "password"} placeholder="Password" {...field} />
+                                            <div className='items-top flex space-x-2'>
+                                                <Checkbox id='showPassword' onClick={() => setShowPassword(!showPassword)}> Show Password</Checkbox>
+                                                <label
+                                                htmlFor="showPassword"
+                                                className="text-sm font-medium leading-none peer-disabled:cursor-not-allowed peer-disabled:opacity-70"
+                                                >
+                                                Show Password
+                                                </label>
+                                            </div>
+                                        </div>
+                                    </FormControl>
+                                    <FormMessage />
+                                </FormItem>
+                            )}
+                        />
+                        <Button disabled={isLoading} type="submit">{isLoading && <Loader2 className='mr-2 h-4 w-4 animate-spin' />} Submit</Button>
+                    </form>
+                </Form>
+            </div>
         </div>
     )    
 }
